requirements.txt
streamlit
gspread
google-auth
pandas
st-gsheets-connection
requests
